requirements.txt
pillow>=7.2.0
torch>=1.7
h5py>=2.10.0
numpy>=1.18.5
scikit-learn>=0.23.1
matplotlib>=3.3.1